@@ -2,40 +2,40 @@
   "name": "antwar",
   "version": "0.3.10",
   "dependencies": {
-    "async": "^0.9.0",
-    "autoprefixer-loader": "1.0.0",
+    "async": "^0.9.2",
+    "autoprefixer-loader": "^1.2.0",
     "babel-core": "^5.4.4",
     "babel-loader": "^5.1.3",
     "coffee-loader": "^0.7.2",
-    "commonmark": "^0.17.1",
+    "commonmark": "^0.19.0",
     "cpr": "^0.4.0",
-    "css-loader": "^0.9.1",
-    "es6-promise": "^2.0.1",
+    "css-loader": "^0.12.1",
+    "es6-promise": "^2.1.1",
     "exports-loader": "^0.6.2",
     "expose-loader": "^0.6.0",
-    "extract-text-webpack-plugin": "^0.3.8",
+    "extract-text-webpack-plugin": "^0.8.0",
     "file-loader": "^0.8.1",
     "imports-loader": "^0.6.3",
     "jshint-loader": "^0.8.3",
     "json-loader": "^0.5.1",
-    "lodash": "^3.3.1",
+    "lodash": "^3.9.1",
     "markdown-loader": "^0.1.2",
-    "mkdirp": "^0.5.0",
+    "mkdirp": "^0.5.1",
     "node-libs-browser": "^0.5.0",
     "portfinder": "^0.4.0",
     "promise.prototype.finally": "^1.0.1",
     "raw-loader": "^0.5.1",
-    "react": "^0.12.2",
-    "react-hot-loader": "^1.1.6",
-    "react-router": "^0.12.4",
-    "remove-markdown": "0.0.4",
-    "sass-loader": "^0.4.2",
+    "react": "^0.13.3",
+    "react-hot-loader": "^1.2.7",
+    "react-router": "^0.13.3",
+    "remove-markdown": "^0.0.4",
+    "sass-loader": "^1.0.2",
     "script-loader": "^0.6.1",
-    "style-loader": "^0.8.3",
+    "style-loader": "^0.12.2",
     "url-loader": "^0.5.5",
-    "webpack": "^1.9.4",
+    "webpack": "^1.9.7",
     "webpack-dev-server": "^1.8.2",
-    "yaml-frontmatter-loader": "^0.0.2"
+    "yaml-frontmatter-loader": "^0.0.3"
   },
   "devDependencies": {},
   "main": "./index.js",
